+++ src/MVFC.Image.Domain/Usings.cs
@@ -5,5 +5,6 @@ global using MVFC.Image.Domain.Apis;
 global using MVFC.Image.Domain.Contracts;
 global using MVFC.Image.Shareable.Configs;
 global using MVFC.Image.Shareable.Requests;
+global using MVFC.Image.Shareable.Response;
 global using MVFC.Mediator;
 global using Refit;