+++ src/MVFC.Image.Shareable/Requests/Base/FileBaseRequest.cs
@@ -1,0 +1,8 @@
+namespace MVFC.Image.Shareable.Requests.Base;
+
+public abstract record FileBaseRequest(
+    string FileName,
+    string ContentType,
+    long Size,
+    string Bucket,
+    DateTime UploadedAt);