+++ src/MVFC.Image.Shareable/Requests/FileDeletePublisherRequest.cs
@@ -1,5 +1,5 @@
 namespace MVFC.Image.Shareable.Requests;
 
-public sealed record FileDeleteRequest(
+public sealed record FileDeletePublisherRequest(
     string FileName) :
     ICommand<Result>;