+++ src/MVFC.Image.Shareable/Requests/FileGalleryRequest.cs
@@ -1,0 +1,4 @@
+namespace MVFC.Image.Shareable.Requests;
+
+public sealed record FileGalleryRequest() :
+    ICommand<Result<FileGalleryResponse>>;