+++ src/MVFC.Image.Shareable/Requests/FileUploadRequest.cs
@@ -1,3 +1,8 @@
 namespace MVFC.Image.Shareable.Requests;
 
-public record FileUploadRequest(string FileName, string ContentType, long Length, byte[] Data) : ICommand<Result<string>>;
+public record FileUploadRequest(
+    string FileName,
+    string ContentType,
+    long Length,
+    byte[] Data) :
+    ICommand<Result<string>>;