+++ src/MVFC.Image.Shareable/Requests/PubSubRequest.cs
@@ -1,8 +1,5 @@
 namespace MVFC.Image.Shareable.Requests;
 
-public sealed record PubSubRequest
-{
-    public PubSubMessageRequest Message { get; set; } = default!;
-
-    public string Subscription { get; set; } = string.Empty;
-}
+public sealed record PubSubRequest(
+    PubSubMessageRequest Message,
+    string Subscription) : ICommand;