+++ src/MVFC.Image.Shareable/Usings.cs
@@ -1,2 +1,4 @@
 global using FluentResults;
+global using MVFC.Image.Shareable.Requests.Base;
+global using MVFC.Image.Shareable.Response;
 global using MVFC.Mediator;