+++ src/MVFC.ImageThumbnail.Worker/Program.cs
@@ -1,31 +1,16 @@
 var builder = WebApplication.CreateBuilder(args);
 
-var storage = await new StorageClientBuilder
-{
-    EmulatorDetection = EmulatorDetection.EmulatorOrProduction
-}
-.BuildAsync();
-
-var publisher = await new PublisherClientBuilder
-{
-    EmulatorDetection = EmulatorDetection.EmulatorOrProduction,
-    TopicName = TopicName.FromProjectTopic("local-project", "thumbnail-created-topic")
-}
-.BuildAsync();
-
-builder.Services.AddSingleton(storage);
-builder.Services.AddSingleton(publisher);
-builder.Services.AddScoped<ThumbnailService>();
+await builder.Services.RegisterThumbnailServicesAsync(builder.Configuration);
 
 var app = builder.Build();
 
-app.MapPost("/pubsub/push", async (PubSubRequest request, ThumbnailService service, CancellationToken ct) =>
+app.MapPost("/pubsub/push", async (PubSubRequest request, IMediator mediator, CancellationToken ct) =>
 {
     var bytes = Convert.FromBase64String(request.Message.Data);
     var json = Encoding.UTF8.GetString(bytes);
-    var evt = JsonSerializer.Deserialize<FileUploadedRequest>(json)!;
+    var evt = JsonSerializer.Deserialize<FileThumbnailRequest>(json)!;
 
-    await service.ProcessAsync(evt, ct);
+    await mediator.Send<FileThumbnailRequest, Result>(evt, ct);
 
     return Results.Ok();
 });